+++ OneclickDonation/build.gradle
@@ -30,11 +30,6 @@ dependencies {
 
 	implementation 'org.xerial:sqlite-jdbc:3.41.2.2'
 	runtimeOnly 'org.hibernate.orm:hibernate-community-dialects:6.2.4.Final'
-
-	//oauth2
-	implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
-
-
 	// jjwt
 	implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
 	runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.11.5'
