+++ OneclickDonation/src/main/java/com/example/OneclickDonation/admin/AdminController.java
@@ -1,10 +1,11 @@
 package com.example.OneclickDonation.admin;
 
 import com.example.OneclickDonation.admin.dto.UpgradeAdminDto;
+import com.example.OneclickDonation.member.dto.MemberDto;
 import com.example.OneclickDonation.post.dto.PostDto;
 import com.example.OneclickDonation.post.service.PostService;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -15,17 +16,13 @@ import org.springframework.web.bind.annotation.*;
 
 @Slf4j
 @Controller
+@RequestMapping("admin")
+@RequiredArgsConstructor
 public class AdminController {
     private final AdminService service;
     private final PostService postService;
 
-    @Autowired
-    public AdminController(AdminService service, PostService postService) {
-        this.service = service;
-        this.postService = postService;
-    }
-
-    @GetMapping("/admin")
+    @GetMapping
     public String adminPage(
             Pageable pageable,
             Model model
@@ -35,7 +32,7 @@ public class AdminController {
         return "admin/home";
     }
 
-    @GetMapping("/admin/upgrades")
+    @GetMapping("/upgrades")
     public String upgradeRequests(
             Pageable pageable,
             Model model
@@ -43,5 +40,25 @@ public class AdminController {
         Page<UpgradeAdminDto> upgradeList = service.listRequests(pageable);
         model.addAttribute("upgradeList", upgradeList);
         return "admin/upgrade";
+    }
+
+    // 수락
+    @PostMapping("/upgrades/{id}/accept")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public UpgradeAdminDto approve(
+            @PathVariable("id")
+            Long id
+    ) {
+        return service.acceptUpgrade(id);
+    }
+
+    // 거절
+    @DeleteMapping("/upgrades/{id}/accept")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public UpgradeAdminDto disapprove(
+            @PathVariable("id")
+            Long id
+    ) {
+        return service.rejectUpgrade(id);
     }
 }
