+++ OneclickDonation/src/main/resources/templates/member/signin.html
@@ -190,8 +190,6 @@
         <p>OR</p>
     </div>
     <!--  소셜 로그인 할 수 있도록 넣기-->
-    <a href="/oauth2/authorization/naver" class="naver-login-btn">네이버</a>
-    <a href="/oauth2/authorization/kakao" class="kakao-login-btn">카카오</a>
 
 </div>
