+++ OneclickDonation/src/main/resources/templates/post/view.html
@@ -113,39 +113,6 @@
           font-size: 14px;
           color: #777;
       }
-
-      .progress-bar {
-          width: 100%;
-          background-color: #f0f0f0;
-          border-radius: 5px;
-          overflow: hidden;
-          height: 20px; /* 높이 설정 */
-      }
-
-      .progress-bar-fill {
-          height: 100%;
-          background-color: #4CAF50;
-          text-align: center;
-          line-height: 20px; /* 세로 가운데 정렬 */
-          color: white;
-          transition: width 0.5s ease-in-out; /* 진행 애니메이션 추가 */
-          position: relative; /* 부모 요소에 대한 상대 위치 설정 */
-      }
-
-      /* 퍼센트 수치 스타일 */
-      .progress-percent {
-          position: absolute; /* 상대 위치 설정 */
-          top: 50%; /* 상단에서 50% 위치 */
-          right: 0; /* 오른쪽에서 시작 */
-          transform: translateY(-50%); /* 세로 중앙 정렬 */
-          width: auto; /* 너비 자동 설정 */
-          color: black; /* 퍼센트 색상 설정 */
-      }
-
-      /* 마우스 호버시 배경색 변경 */
-      .progress-bar-fill:hover {
-          background-color: #45a049;
-      }
   </style>
 </head>
 <body>
@@ -175,13 +142,6 @@
     <div>
       <strong>목표 모금액:</strong> <br>
       <span th:text="${post.targetAmount + '원'}"></span>
-    </div>
-  </div>
-  <br>
-  <div class="container">
-    <div class="progress-bar" id="progress-bar">
-      <div class="progress-bar-fill" id="progress-bar-fill"></div>
-      <div class="progress-percent" id="progress-percent"></div>
     </div>
   </div>
   <br>
@@ -221,42 +181,6 @@
     <input type="submit" value="댓글 작성"> <br>
   </form>
 </div>
-
-
-<script>
-    // updateProgressBar 함수를 정의합니다.
-    function updateProgressBar(supportAmount, targetAmount) {
-        // 현재 모금액과 목표 모금액을 가져와서 계산합니다.
-        var percent = supportAmount >= targetAmount ? 100 : (supportAmount / targetAmount) * 100;
-        // 게이지 바를 채우는 요소를 가져옵니다.
-        var progressBarFill = document.getElementById("progress-bar-fill");
-        // 게이지 바를 채우는 요소의 너비를 퍼센트에 따라 조정합니다.
-        progressBarFill.style.width = percent + "%";
-        progressBarFill.innerText = percent.toFixed(2) + "% (" + supportAmount + "원" + ")";
-    }
-
-    // 페이지가 로드될 때 서버에서 데이터를 가져와서 게이지 바를 업데이트합니다.
-    document.addEventListener("DOMContentLoaded", function() {
-        // AJAX 요청을 생성합니다.
-        var xhr = new XMLHttpRequest();
-        // 현재 페이지 URL에서 게시물 ID를 가져옵니다.
-        var postId = window.location.pathname.split("/")[2]; // /post/{id}/support-amount-target-amount 에서 {id}를 가져옵니다.
-        // GET 메서드를 사용하여 /{id}/support-amount-target-amount 엔드포인트로 요청을 보냅니다.
-        xhr.open("GET", "/post/" + postId + "/support-amount-target-amount", true);
-        // 응답이 도착하면 이벤트 리스너를 실행합니다.
-        xhr.onload = function () {
-            // 만약 요청이 성공했다면
-            if (xhr.status === 200) {
-                // 서버에서 받은 JSON 데이터를 JavaScript 객체로 변환합니다.
-                var data = JSON.parse(xhr.responseText);
-                // 가져온 데이터를 이용하여 게이지 바를 업데이트합니다.
-                updateProgressBar(data.supportAmount, data.targetAmount);
-            }
-        };
-        // AJAX 요청을 보냅니다.
-        xhr.send();
-    });
-</script>
 
 </body>
 </html>
